requirements.txt
uvicorn==0.23.2
python-dotenv==1.0.1
fastapi==0.103.2
langchain==0.3.18
langchain-community==0.3.17
streamlit==1.37.1
